+++ internal/component.go
@@ -34,10 +34,11 @@
 package internal
 
 import (
-	"blobfuse2/common"
-	"blobfuse2/internal/handlemap"
 	"context"
 	"syscall"
+
+	"github.com/Azure/azure-storage-fuse/v2/common"
+	"github.com/Azure/azure-storage-fuse/v2/internal/handlemap"
 )
 
 type ComponentPriority int
@@ -69,7 +70,7 @@ type Component interface {
 	// Pipeline participation related methods
 	Name() string
 	SetName(string)
-	Configure() error
+	Configure(bool) error
 	Priority() ComponentPriority
 
 	SetNextComponent(c Component)